+++ changelog.md
@@ -51,6 +51,7 @@ Codion Change Log
 - EntitySearchField refactored.
 - EntitySearchField.Build.separator() and stringFactory() added, related refactoring.
 - EntitySearchField.Builder.selectAllOnFocusGained() removed, related refactoring.
+- EntitySearchField, EntitySearchFieldPanel.MultiSelectionBuilder and SingleSelectionBuilder added along with Factory, related refactoring.
 ### is.codion.tools.generator.ui
 - DomainGeneratorPanel bug fixed, did not display exceptions happening during async table model refresh.